Address comments, rename ip files
